+++ packages/utils/src/functions/index.ts
@@ -5,7 +5,6 @@ export * from "./chunk";
 export * from "./cn";
 export * from "./combine-words";
 export * from "./construct-metadata";
-export * from "./cents";
 export * from "./currency-formatter";
 export * from "./currency-zero-decimal";
 export * from "./datetime";
@@ -38,6 +37,7 @@ export * from "./smart-truncate";
 export * from "./stable-sort";
 export * from "./text-fetcher";
 export * from "./time-ago";
+export * from "./to-cents-number";
 export * from "./trim";
 export * from "./truncate";
 export * from "./urls";
